refactor: move network naming rules into cella-backend
repo_network_name / workspace_network_name / CELLA_NETWORK_NAME and
the managed-network labels are backend-agnostic identity rules, not
Docker specifics. Hoist them next to ManagedNetwork so the Apple
Container backend can derive identical network names; cella-docker
re-exports them unchanged.
